[CDAP-19408] Add job watchers in non-system k8s namespaces
